# Improve CI performance by limiting max workers