@wchargin 2018-09-21
:heart: ing this PR because it appears to have resolved the spurious
5000ms flaky timeouts, saving me much frustration. Yay!